qa_backend_apifinal/qa_backend_apifinal/requirements.txt
fastapi
uvicorn
transformers
torch
scikit-learn
numpy
python-docx
pymupdf
